Test ServerConnection with all nodelay settings
Same test file change as seen in cro-core, except for tls.t instead of tcp.t,
since (in this section at least) the TLS test file is nearly identical to the
TCP one.

From the original commit message:

The original tests in this section can easily be reused, so just run them
three times: without the nodelay option, with it off, and with it on.